requirements.txt
asgiref==3.9.1
Django==5.2.5
django-ipware==7.0.1
python-ipware==3.0.0
sqlparse==0.5.3
